Redux-toolkit setup
Initial redux-toolkit setup with example counter component. This counter component and related folders will be deleted after our team meeting on 23rd of July 2024. We can wait with merging this branch on development until 24rd of July. I will let the team know, when this branch is ready to merge with development. Thank you :-)